Make it 64 bits compatible
	* include/libport/pod-cast.hh, lib/libport/symbol.cc: Use types
	large enough for 64 bits systems.

+++ include/libport/pod-cast.hh
@@ -15,7 +15,7 @@ namespace libport
   /// for instance in order to embed them into a union.
   ///
   /// Of course, no ctor/dtor: the whole point is to get a POD.
-  template <typename T, typename U = unsigned>
+  template <typename T, typename U = long long>
   class pod_caster
   {
   public:
